// types.ts
export interface ReportData {
  tipoMovimiento: string;
  expediente: string;
  establecimiento: string;
  telefono: string;
  delegacion: string;
  reparticion: string;
  cuil: string;
  rol: string;
  apellidoYNombre: string;
  situacionDeRevista: string;
  fecha: string;
  cc: string;
  cargoACubrir: string;
  concurso: string;
  resolucion: string;
  // reemplazaA?: string;
  // motivoDeCese?: string;
}